Reorganized solution
