+++ Potmobile/Assets.cs
@@ -7,13 +7,12 @@ namespace Potmobile
     public class Assets
     {
         public static AssetBundle assetBundle;
-        public static Mesh jeepMesh;
 
         internal static string assemblyDir
         {
             get
             {
-                return System.IO.Path.GetDirectoryName(Potmobile.pluginInfo.Location);
+                return System.IO.Path.GetDirectoryName(PotmobilePlugin.pluginInfo.Location);
             }
         }
 
@@ -31,12 +30,6 @@ namespace Potmobile
                 bankStream.Read(bytes, 0, bytes.Length);
                 R2API.SoundAPI.SoundBanks.Add(bytes);
             }
-            LoadAssets();
-        }
-
-        private static void LoadAssets()
-        {
-            jeepMesh = Assets.assetBundle.LoadAsset<Mesh>("Assets/Resources/mdlHaulerJeep.fbx");
         }
     }
 }
