+++ PotmobileUnity/AssetBundles/StandaloneWindows/StandaloneWindows.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 3848024072
+CRC: 3353402987
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
